% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reports.R
\name{get_reports}
\alias{get_reports}
\title{Download World Malaria Report PDFs}
\usage{
get_reports(years = 2017:2024, outdir = ".")
}
\arguments{
\item{years}{which years to download pdfs}

\item{outdir}{the path to the directory where pdfs should be downloaded}
}
\description{
Download World Malaria Report PDFs
}
\examples{
get_reports()
}
