% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPDFs.R
\name{getWMRPDFs}
\alias{getWMRPDFs}
\title{Download World Malaria Report PDFs}
\usage{
getWMRPDFs(years = 2017:2023, outdir = ".")
}
\arguments{
\item{years}{which years to download pdfs}

\item{outdir}{the path to the directory where pdfs should be downloaded}
}
\description{
Download World Malaria Report PDFs
}
\examples{
getWMRPDFs()
}
