% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_make_numeric.R
\name{try_make_numeric}
\alias{try_make_numeric}
\title{Try to convert columns to numeric, replacing any na values with NA}
\usage{
try_make_numeric(
  df,
  cols,
  na = c("-", ".", ""),
  str_remove = NULL,
  warn = TRUE,
  stopAtFirstError = TRUE
)
}
\arguments{
\item{df}{the \code{data.frame} to use}

\item{cols}{the columns to convert to numeric, either as a vector of column names or a vector of column numbers}

\item{na}{the values to replace with NA, default is c("-", ".", "")}

\item{str_remove}{a regular expression to remove from the strings after replacing with NA values but before converting to numeric}

\item{warn}{if TRUE (the default), then a warning is issued if a column cannot be converted to numeric}

\item{stopAtFirstError}{if TRUE (the default), then the function stops at the first error, otherwise it continues}
}
\value{
the \code{data.frame} with the columns converted to numeric
}
\description{
Try to convert columns to numeric, replacing any na values with NA
}
\keyword{internal}
