% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wmr2018}
\alias{wmr2018}
\title{World Health Organization World Malaria Report 2018}
\format{
A list containing a tibble for each item in the annex:
\describe{
\item{wmr2018a}{p98: Policy adoption, 2017. Tanzania dropped.}
\item{wmr2018b}{p102: Antimalarial drug policy, 2017.}
\item{wmr2018c}{p104: Funding for malaria control, 2015–2017.}
\item{wmr2018d}{p116: Commodities distribution and coverage, 2015–2017.}
\item{wmr2018e}{p122: Household survey results, 2015–2017. No WHO Region column.}
\item{wmr2018f}{p124: Population at risk and estimated malaria cases and deaths, 2010–2017.}
\item{wmr2018g}{p138: Population at risk and reported malaria cases by place of care, 2017. Be careful not to double count Tanzania which is split into Mainland and Zanzibar.}
\item{wmr2018h}{p142: Reported malaria cases by method of confirmation, 2010–2017}
\item{wmr2018i}{p154: Reported malaria cases by species, 2010–2017}
\item{wmr2018j}{p164: Reported malaria deaths, 2010–2017}
}
}
\source{
Report: \url{https://iris.who.int/bitstream/handle/10665/275867/9789241565653-eng.pdf?sequence=1}

Annex: \url{https://cdn.who.int/media/docs/default-source/malaria/world-malaria-reports/wmr2018-excel-annexes.zip}
}
\usage{
wmr2018
}
\description{
Annex data from the World Health Organization World Malaria Report 2018
Note that Tanzania is typically split into Mainland and Zanzibar
}
\keyword{datasets}
