% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_match_dict.R
\name{case_match_dict}
\alias{case_match_dict}
\title{Helper function useful when mapping values in character strings}
\usage{
case_match_dict(.x, dict, trimFirst = TRUE)
}
\arguments{
\item{.x}{a vector to match against}

\item{dict}{a named list - the names should correspond to the values in .x
and the values should be the new values}
}
\description{
Helper function useful when mapping values in character strings
}
\examples{
df |> dplyr::mutate(Footnotes=case_match_dict(Footnotes, dict=c("2"="footnote1", "5,6"="footnote 5. footnote6."))
}
