% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_whowmr_reports.R
\name{download_whowmr_reports}
\alias{download_whowmr_reports}
\title{Download WHO World Malaria Report PDFs}
\usage{
download_whowmr_reports(years = 2017:2025, outdir = ".")
}
\arguments{
\item{years}{which years to download pdfs}

\item{outdir}{the path to the directory where pdfs should be downloaded}
}
\description{
Download WHO World Malaria Report PDFs
}
\examples{
download_whowmr_reports()
}
