% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_annex_dirs_exist.R
\name{ensure_annex_dir_exists}
\alias{ensure_annex_dir_exists}
\title{Ensure the unzipped annex exists for the given year(s)}
\usage{
ensure_annex_dir_exists(url, filename, removeZip = FALSE)
}
\arguments{
\item{url:}{The URL to download the annex from}

\item{filename:}{The filename of the zip file to be saved in
data-raw/downloads and unzipped as a directory in
data-raw/unzipped}

\item{removeZip:}{Whether to remove the zip file after unzipping}
}
\value{
named list of files in the unzipped annex
}
\description{
Ensure the unzipped annex exists for the given year(s)
}
