% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_annex_dirs_exist.R
\name{ensure_annex_dirs_exist}
\alias{ensure_annex_dirs_exist}
\title{Ensure the unzipped annex exists for the given year(s)}
\usage{
ensure_annex_dirs_exist(years = 2017:2023)
}
\arguments{
\item{years:}{A numeric vector of years to download and unzip the annexes for}
}
\value{
fileTree: A named list of all files in the unzipped annexes, with the
names of the list being the directory of the unzipped annex
and the values being the file paths relative to that path.
}
\description{
Ensure the unzipped annex exists for the given year(s)
}
