% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_who_region.R
\name{split_who_region}
\alias{split_who_region}
\title{Split the contents of \verb{WHO region\\\\nCountry/area} column in an annex sheet.}
\usage{
split_who_region(
  df,
  col_region_area = 1,
  region_name = "WHO Region",
  area_name = "Country/area",
  fill_down_first = TRUE
)
}
\arguments{
\item{df}{the dataframe from the excel sheet in the WMR annexes}

\item{col_region_area}{the number/name for the column containing the \verb{WHO region\\\\nCountry/area} column}

\item{region_name}{the name of the new region column pulled out from the \verb{WHO region\\\\nCountry/area} column}

\item{area_name}{the name of the new country/area column pulled out from the \verb{WHO region\\\\nCountry/area} column}

\item{fill_down_first}{if TRUE (the default), fill down the region/area column before splitting the region and area names.

I highly recommend using the \code{fill_down_first} option.}
}
\value{
the dataframe with the \code{col_region_area} column split into \code{region_name} and \code{area_name} columns.
}
\description{
Split the contents of \verb{WHO region\\\\nCountry/area} column in an annex sheet.
}
\examples{
wmr2017a <- readxl::read_excel("wmr2017-excel-annexes/wmr2017-annex-table-a.xls", range = "A2:Q102")
split_who_region(wmr2017a, col_region_area=1, region_name="WHO Region", area_name="Country/area")
}
