% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not_installed}
\alias{stop_if_not_installed}
\title{Stop if suggested packages are not installed}
\usage{
stop_if_not_installed(...)
}
\arguments{
\item{pkgs}{a vector of package names to check if they are installed}
}
\description{
Stop if suggested packages are not installed
}
\examples{
stop_if_not_installed(c("readxl", "dplyr"))
stop_if_not_installed()
}
